Refactor com_atproto::label
