+++ atprotolib-rs/src/types/com_atproto/label/api_responses.rs
@@ -2,10 +2,6 @@ use serde::{Deserialize, Serialize};
 
 use super::Label;
 
-/*
-    com.atproto.label.queryLabels
-*/
-
 /// Represents a response to a request to query labels.
 #[derive(Serialize, Deserialize, Debug)]
 pub struct QueryLabelsResponse {
